requirements.txt
sqlalchemy==2.0.42
fastapi==0.116.1
pydantic-settings==2.10.1
pydantic==2.11.7
starlette==0.47.2
config==0.5.1
alembic==1.16.4
python-dotenv==1.1.1
asyncpg==0.30.0
uvicorn==0.35.0